requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.20
pypdf==3.17.4
python-docx==1.1.0
openai==1.6.1
tiktoken==0.5.2
sentence-transformers==2.2.2
streamlit-chat==0.1.1
python-dotenv==1.0.0